requirements.txt
fastapi
uvicorn
smpplib
python-dotenv
xmltodict
requests>=2.25.1
pymongo
python-multipart
aiohttp


